[core] Moved ony pickle_dumps into _common
