+++ python/ray/remote_function.py
@@ -10,12 +10,12 @@ import ray._common.signature
 from ray import Language, cross_language
 from ray._common import ray_option_utils
 from ray._common.ray_option_utils import _warn_if_using_deprecated_placement_group
+from ray._common.serialization import pickle_dumps
 from ray._private.auto_init_hook import wrap_auto_init
 from ray._private.client_mode_hook import (
     client_mode_convert_function,
     client_mode_should_convert,
 )
-from ray._private.serialization import pickle_dumps
 from ray._private.utils import get_runtime_env_info, parse_runtime_env_for_task_or_actor
 from ray._raylet import (
     STREAMING_GENERATOR_RETURN,
